+++ pkg/analysis_server/lib/src/lsp/constants.dart
@@ -149,6 +149,7 @@ abstract final class CustomMethods {
   static const publishFlutterOutline = Method(
     'dart/textDocument/publishFlutterOutline',
   );
+  static const summary = Method('dart/textDocument/summary');
   static const super_ = Method('dart/textDocument/super');
   static const imports = Method('dart/textDocument/imports');
   static const dartTextDocumentContent = Method('dart/textDocumentContent');