+++ pkg/analysis_server/test/lsp/test_all.dart
@@ -58,6 +58,7 @@ import 'server_test.dart' as server;
 import 'signature_help_test.dart' as signature_help;
 import 'snippets_test.dart' as snippets;
 import 'source_edits_test.dart' as source_edits;
+import 'summary_test.dart' as summary;
 import 'super_test.dart' as get_super;
 import 'temporary_overlay_operation_test.dart' as temporary_overlay_operation;
 import 'type_definition_test.dart' as type_definition;
@@ -102,6 +103,7 @@ void main() {
     flutter_outline.main();
     folding.main();
     format.main();
+    summary.main();
     get_super.main();
     hover.main();
     implementation.main();