+++ runtime/vm/bytecode_reader.h
@@ -23,7 +23,8 @@ class BytecodeLoader {
   BytecodeLoader(Thread* thread, const TypedDataBase& binary);
   ~BytecodeLoader();
 
-  FunctionPtr LoadBytecode();
+  FunctionPtr LoadBytecode(bool load_code = true);
+  void LoadPendingCode();
 
   TypedDataBasePtr binary() const { return binary_.ptr(); }
   ArrayPtr bytecode_component_array() const {
@@ -43,6 +44,7 @@ class BytecodeLoader {
   Thread* thread_;
   const TypedDataBase& binary_;
   Array& bytecode_component_array_;
+  const GrowableObjectArray& pending_classes_;
   Array& bytecode_offsets_map_;
 
   DISALLOW_COPY_AND_ASSIGN(BytecodeLoader);
@@ -209,7 +211,10 @@ class BytecodeReaderHelper : public ValueObject {
   void ReadClassDeclaration(const Class& cls);
   void ReadLibraryDeclaration(const Library& library,
                               const GrowableObjectArray& pending_classes);
-  void ReadLibraryDeclarations(intptr_t num_libraries);
+  void ReadLibraryDeclarations(intptr_t num_libraries,
+                               const GrowableObjectArray& pending_classes,
+                               bool load_code);
+  void ReadPendingCode(const GrowableObjectArray& pending_classes);
   void FindModifiedLibraries(BitVector* modified_libs, intptr_t num_libraries);
 
   LibraryPtr ReadMain();